+++ src/clojure/lazytest/context.clj
@@ -15,39 +15,39 @@
 
 (defn run-befores
   [obj]
-  (doseq [before-fn (-> obj :context :before)
+  (doseq [before-fn (-> obj :lazytest/context :before)
           :when (fn? before-fn)]
     (before-fn)))
 
 (defn run-before-eachs
   [obj]
-  (doseq [before-each-fn (-> obj :context :before-each)
+  (doseq [before-each-fn (-> obj :lazytest/context :before-each)
           :when (fn? before-each-fn)]
     (before-each-fn)))
 
 (defn run-after-eachs
   [obj]
-  (doseq [after-each-fn (-> obj :context :after-each)
+  (doseq [after-each-fn (-> obj :lazytest/context :after-each)
           :when (fn? after-each-fn)]
     (after-each-fn)))
 
 (defn run-afters
   [obj]
-  (doseq [after-fn (-> obj :context :after)
+  (doseq [after-fn (-> obj :lazytest/context :after)
           :when (fn? after-fn)]
     (after-fn)))
 
 (defn combine-arounds
   [obj]
-  (when-let [arounds (-> obj :context :around seq)]
+  (when-let [arounds (-> obj :lazytest/context :around seq)]
     (c.t/join-fixtures arounds)))
 
 (defn propagate-eachs
   [parent-meta child]
   (let [child-meta (meta child)
         updated-meta (-> child-meta
-                         (assoc-in [:context :before-each] (into (vec (-> parent-meta :context :before-each))
-                                                                 (-> child-meta :context :before-each)))
-                         (assoc-in [:context :after-each] (into (vec (-> child-meta :context :after-each))
-                                                                (-> parent-meta :context :after-each))))]
+                         (assoc-in [:lazytest/context :before-each] (into (vec (-> parent-meta :lazytest/context :before-each))
+                                                                 (-> child-meta :lazytest/context :before-each)))
+                         (assoc-in [:lazytest/context :after-each] (into (vec (-> child-meta :lazytest/context :after-each))
+                                                                (-> parent-meta :lazytest/context :after-each))))]
     (with-meta child updated-meta)))